Migrate jira access plugin tests
