+++ integrations/access/jira/client.go
@@ -42,9 +42,10 @@ const (
 
 	jiraMaxConns    = 100
 	jiraHTTPTimeout = 10 * time.Second
-	// Teleport has a 4096 character limit for the reason field so we
+	// JiraReasonLimit is the max size for a reason in jira.
+	// Teleport has a 4096-character limit for the reason field, so we
 	// truncate all reasons to a generous but conservative limit
-	jiraReasonLimit = 3000
+	JiraReasonLimit = 3000
 
 	jiraStatusUpdateTimeout time.Duration = 10 * time.Second
 )
@@ -450,11 +451,11 @@ func (j *Jira) AddResolutionComment(ctx context.Context, id string, resolution R
 }
 
 func truncateReasonFields(reqData RequestData) RequestData {
-	if reqData.Resolution.Reason != "" && len(reqData.Resolution.Reason) > jiraReasonLimit {
-		reqData.Resolution.Reason = reqData.Resolution.Reason[:jiraReasonLimit]
+	if reqData.Resolution.Reason != "" && len(reqData.Resolution.Reason) > JiraReasonLimit {
+		reqData.Resolution.Reason = reqData.Resolution.Reason[:JiraReasonLimit]
 	}
-	if reqData.RequestReason != "" && len(reqData.RequestReason) > jiraReasonLimit {
-		reqData.RequestReason = reqData.RequestReason[:jiraReasonLimit]
+	if reqData.RequestReason != "" && len(reqData.RequestReason) > JiraReasonLimit {
+		reqData.RequestReason = reqData.RequestReason[:JiraReasonLimit]
 	}
 	return reqData
 }
